refactor: move names out of init.
