+++ metriq_gym/run.py
@@ -18,7 +18,7 @@ from qbraid.runtime import (
     load_provider,
 )
 
-from metriq_gym.benchmarks import (
+from metriq_gym.config import (
     BENCHMARK_DATA_CLASSES,
     BENCHMARK_HANDLERS,
     get_available_benchmarks,
@@ -31,7 +31,7 @@ from metriq_gym.exporters.json_exporter import JsonExporter
 from metriq_gym.job_manager import JobManager, MetriqGymJob
 from metriq_gym.qplatform.job import job_status
 from metriq_gym.schema_validator import load_and_validate, validate_and_create_model
-from metriq_gym.benchmarks import JobType
+from metriq_gym.config import JobType
 
 logging.basicConfig(level=logging.WARNING)
 logger = logging.getLogger("metriq_gym")
